+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a new constructor for `GaussLegendre` from an index range describing the cell edges.
+- Add a `GradientCreator` operator to group derivative calculations.
 
 ### Fixed